Testing

+++ bradenasmith/Controllers/CommentsController.cs
@@ -3,6 +3,7 @@ using bradenasmith.Models;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
 using System.Text.Json;
+using Serilog;
 
 namespace bradenasmith.Controllers
 {
@@ -19,45 +20,32 @@ namespace bradenasmith.Controllers
         [Route("/Blogs/{topic}/Comment/New")]
         public IActionResult Index(string topic, Comment comment, int BlogPostId)//BlogPostId comes from a HIDDEN input.
         {
-            if (Request.Cookies["AnonUser"] == null)
+            if(topic != null && comment != null && BlogPostId != null) 
             {
-                string anonId = Guid.NewGuid().ToString();
-                var anonUser = new User();
+                if (Request.Cookies["AnonUser"] == null)
+                {
+                    string anonId = Guid.NewGuid().ToString();
+                    var anonUser = new User();
 
-                anonUser.Id = anonId;
-                Response.Cookies.Append("AnonUser", anonId);
-                comment.AnonId = anonId;
-                comment.BlogPost = _context.BlogPosts.Find(BlogPostId);
-                comment.CreatedAt = DateTime.Now.ToUniversalTime();
-                anonUser.Comments.Add(comment);
-            }
-            else
-            {
-                var anonUser = new User();
+                    anonUser.Id = anonId;
+                    Response.Cookies.Append("AnonUser", anonId);
+                    comment.AnonId = anonId;
+                    comment.BlogPost = _context.BlogPosts.Find(BlogPostId);
+                    comment.CreatedAt = DateTime.Now.ToUniversalTime();
+                    anonUser.Comments.Add(comment);
+                }
+                else
+                {
+                    var anonUser = new User();
 
-                anonUser.Id = Request.Cookies["AnonUser"];
+                    anonUser.Id = Request.Cookies["AnonUser"];
 
-                comment.BlogPost = _context.BlogPosts.Find(BlogPostId);
-                comment.CreatedAt = DateTime.Now.ToUniversalTime();
-                comment.AnonId = anonUser.Id;
-                anonUser.Comments.Add(comment);
-            }
-            _context.Comments.Add(comment);
-            _context.SaveChanges();
-
-            return Redirect($"/Blogs/{topic}");
-        }
-
-        [HttpPost]
-        [Route("/Blogs/{topic}/Comment/{commentId}/Edit")]
-        public IActionResult Edit(string topic, int commentId, Comment updatedComment)
-        {
-            if(updatedComment.AnonId == Request.Cookies["AnonUserId"])
-            {
-                var blog = _context.BlogPosts.Include(e => e.Comments).Where(e => e.Topic.ToLower() == topic.ToLower()).FirstOrDefault();
-                var comment = blog.Comments.FirstOrDefault(c => c.Id == commentId);
-
-                comment.Content = updatedComment.Content;
+                    comment.BlogPost = _context.BlogPosts.Find(BlogPostId);
+                    comment.CreatedAt = DateTime.Now.ToUniversalTime();
+                    comment.AnonId = anonUser.Id;
+                    anonUser.Comments.Add(comment);
+                }
+                _context.Comments.Add(comment);
                 _context.SaveChanges();
 
                 return Redirect($"/Blogs/{topic}");
@@ -69,15 +57,65 @@ namespace bradenasmith.Controllers
         }
 
         [HttpPost]
+        [Route("/Blogs/{topic}/Comment/{commentId}/Edit")]
+        public IActionResult Edit(string topic, int commentId, Comment updatedComment)
+        {
+            if(topic != null && commentId != null && updatedComment != null)
+            {
+                if (updatedComment.AnonId == Request.Cookies["AnonUserId"])
+                {
+                    try
+                    {
+                        var blog = _context.BlogPosts.Include(e => e.Comments).Where(e => e.Topic.ToLower() == topic.ToLower()).FirstOrDefault();
+                        var comment = blog.Comments.FirstOrDefault(c => c.Id == commentId);
+
+                        comment.Content = updatedComment.Content;
+                        _context.SaveChanges();
+
+                        return Redirect($"/Blogs/{topic}");
+                    }
+                    catch(Exception ex) 
+                    {
+                        Log.Warning($"Failed to fetch Db data, error: {ex.Message}");
+                        return BadRequest();
+                    }
+                }
+                else
+                {
+                    return BadRequest();
+                }
+            }
+            else
+            {
+                return BadRequest();
+            }
+        }
+
+        [HttpPost]
         [Route("/Blogs/{topic}/Comment/{commentId}/Delete")]
         public IActionResult Delete(string topic, int commentId)
         {
-            var blog = _context.BlogPosts.Include(e => e.Comments).Where(e => e.Topic.ToLower() == topic.ToLower()).FirstOrDefault();
-            var comment = blog.Comments.FirstOrDefault(e => e.Id == commentId);
-            _context.Comments.Remove(comment);
-            _context.SaveChanges();
+            if(topic != null && commentId != null)
+            {
+                try
+                {
 
-            return Redirect($"/Blogs/{topic}");
+                }
+                catch(Exception ex)
+                {
+                    Log.Warning($"Failed to fetch Db data, error: {ex.Message}");
+                }
+                var blog = _context.BlogPosts.Include(e => e.Comments).Where(e => e.Topic.ToLower() == topic.ToLower()).FirstOrDefault();
+                var comment = blog.Comments.FirstOrDefault(e => e.Id == commentId);
+                _context.Comments.Remove(comment);
+                _context.SaveChanges();
+
+                return Redirect($"/Blogs/{topic}");
+            }
+            else
+            {
+                return BadRequest();
+            }
         }
     }
 }
